move thigns around
